Extract TaskExecution to its own class
